# backend/docs/swagger.yaml
definitions:
  entity.Asset:
    properties:
      amount:
        example: 1000000
        type: integer
      asset_type:
        type: string
      code:
        example: USDC
        type: string
      distributor:
        $ref: '#/definitions/entity.Wallet'
      id:
        example: 1
        type: integer
      image:
        items:
          type: integer
        type: array
      issuer:
        $ref: '#/definitions/entity.Wallet'
      name:
        example: USD Coin
        type: string
    type: object
  entity.Contract:
    properties:
      address:
        type: string
      asset:
        $ref: '#/definitions/entity.Asset'
      created_at:
        type: string
      id:
        example: 1
        type: integer
      min_deposit:
        type: integer
      name:
        example: Smart Contract
        type: string
      penalty_rate:
        type: integer
      term:
        type: integer
      vault:
        $ref: '#/definitions/entity.Vault'
      yield_rate:
        type: integer
    type: object
...
  entity.Key:
    properties:
      id:
        example: 1
        type: integer
      publicKey:
        example: GCK...
        type: string
      walletId:
        example: 1
        type: integer
      weight:
        example: 3
        type: integer
    type: object
...
  entity.Role:
    properties:
      admin:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Admin
        type: string
    type: object
  entity.RoleDelete:
    properties:
      id:
        example: 1
        type: integer
      new_users_role_id:
        example: 1
        type: integer
    type: object
...
  entity.RoleRequest:
    properties:
      name:
        example: Admin
        type: string
    type: object
...
  entity.Vault:
    properties:
      active:
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Treasury
        type: string
      vault_category:
        $ref: '#/definitions/entity.VaultCategory'
      wallet:
        $ref: '#/definitions/entity.Wallet'
    type: object
  entity.VaultCategory:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Treasury
        type: string
      theme:
        example: blue
        type: string
    type: object
  entity.Wallet:
    properties:
      funded:
        type: boolean
      id:
        example: 1
        type: integer
      key:
        $ref: '#/definitions/entity.Key'
      type:
        example: sponsor
        type: string
    type: object
...
  v1.CreateContractRequest:
    properties:
      address:
        example: GSDSC...
        type: string
      asset_id:
        example: "1"
        type: string
      min_deposit:
        example: 1
        type: integer
      name:
        example: Treasury
        type: string
      penalty_rate:
        example: 1
        type: integer
      term:
        example: 1
        type: integer
      vault_id:
        example: "1"
        type: string
      yield_rate:
        example: 1
        type: integer
    required:
    - address
    - asset_id
    - min_deposit
    - name
    - penalty_rate
    - term
    - vault_id
    - yield_rate
    type: object
...
  v1.response:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /contract:
    post:
      consumes:
      - application/json
      description: Create new contract
      parameters:
      - description: Contract info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateContractRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Contract'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Create a new contract
      tags:
      - Contract
  /contracts:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all contracts, with optional pagination
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items per page for pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Contract'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get all contracts
      tags:
      - Contract
...
  /role:
    get:
      consumes:
      - application/json
      description: List role
      parameters:
      - description: Type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Role'
            type: array
      summary: List
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Role info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.RoleRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Create a new role
      tags:
      - Role
...
  /role/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role by providing the Role ID and the updated information.
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Role info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated role information
          schema:
            $ref: '#/definitions/entity.RoleDelete'
        "400":
          description: 'Bad Request: Invalid input data'
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: 'Not Found: Role not found'
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: 'Internal Server Error: Failed to delete role'
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete a role
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update a role by providing the Role ID and the updated information.
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Role info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.RoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated role information
          schema:
            $ref: '#/definitions/entity.Role'
        "400":
          description: 'Bad Request: Invalid input data'
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: 'Not Found: Role not found'
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: 'Internal Server Error: Failed to update role'
          schema:
            $ref: '#/definitions/v1.response'
      summary: Update a role
      tags:
      - Role
...
swagger: "2.0"